Fix cleave urlencode to cope with characters with bit 7 set.
In this line, pstr is a signed char.

   *pret++ = to_hex(*pstr >> 4);

pstr is sign extended to an int before shifting. So if pstr
is negative it becomes very very negative, then to_hex runs
off the end of it's array and reads uninitialized memory.

While here, beef up the logging in popen3 so that next time
this happens we can catch the odd arguments being passed.