Use offsetof() from <stddef.h> instead of our own macro
Building with clang 3.4 fails because it complains that we redefine
the offsetof() macro, because <stddef.h> is included through some
implicit path.  Since offsetof() is a standard part of <stddef.h> at
least since C99, let's switch to including <stddef.h> explicitly and
delete of our local version of offsetof().

Signed-off-by: Roland Dreier <roland@purestorage.com>